pyproject.toml
[tool.poetry]
name = "certify"
version = "1.0.0"
description = "Certificate Verification System API"
authors = ["Dote"]
license = "GNU GENERAL PUBLIC LICENSE V3"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.70.1"
uvicorn = "^0.16.0"
google-cloud-firestore = "^2.3.4"
loguru = "^0.5.3"
python-jose = "^3.3.0"
cryptography = "^36.0.1"
passlib = "^1.7.4"
weasyprint = "^54.0"
Jinja2 = "^3.0.3"
Pillow = "^9.0.0"
google-cloud-storage = "^1.44.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
